codacy based formatting

+++ src/scene/camera.py
@@ -19,7 +19,7 @@ class Camera:
         """ Generate the pixel locations on the camera plane. """
 
         width, height = self.resolution
-        AR = width / height
+        AR = int(width / height)
         yy, zz = np.mgrid[AR:-AR:complex(0, width), 1:-1:complex(0, height)]
         xx = np.ones(shape=yy.shape) * (1 / np.tan(np.radians(self.field_of_view) / 2))  # Distance of grid from origin
